extract Converter from common_nodes
