+++ bioimageio/spec/_internal/common_nodes.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 import collections.abc
 import traceback
-from abc import ABC, abstractmethod
+from abc import ABC
 from copy import deepcopy
 from io import BytesIO
 from pathlib import Path
@@ -12,16 +12,13 @@ from typing import (
     Any,
     ClassVar,
     Dict,
-    Final,
     FrozenSet,
-    Generic,
     List,
     Optional,
     Protocol,
     Tuple,
     Type,
     Union,
-    cast,
     get_type_hints,
 )
 from zipfile import ZipFile
@@ -37,14 +34,7 @@ from pydantic import (
     model_validator,
 )
 from pydantic_core import PydanticUndefined, core_schema
-from typing_extensions import (
-    Annotated,
-    LiteralString,
-    Self,
-    TypeVar,
-    TypeVarTuple,
-    Unpack,
-)
+from typing_extensions import Annotated, LiteralString, Self
 
 from ..summary import (
     WARNING_LEVEL_TO_NAME,
@@ -63,7 +53,6 @@ from .packaging_context import PackagingContext
 from .root_url import RootHttpUrl
 from .url import HttpUrl
 from .utils import (
-    assert_all_params_set_explicitly,
     get_format_version_tuple,
 )
 from .validation_context import (
@@ -152,88 +141,6 @@ class StringNode(collections.UserString, ABC):
             _ = self._node.model_dump(mode="json")
 
         return self.data
-
-
-SRC = TypeVar("SRC", bound=Union[Node, StringNode])
-TGT = TypeVar("TGT", bound=Node)
-
-
-# converter without any additional args or kwargs:
-# class Converter(Generic[SRC, TGT], ABC):
-#     # src: ClassVar[Type[SRC]]
-#     # tgt: ClassVar[Type[TGT]]
-#     # note: the above is not yet possible, see https://github.com/python/typing/discussions/1424
-#     # we therefore use an instance
-#     def __init__(self, src: Type[SRC], tgt: Type[TGT], /):
-#         super().__init__()
-#         self.src: Final[Type[SRC]] = src
-#         self.tgt: Final[Type[TGT]] = tgt
-
-#     @abstractmethod
-#     def _convert(self, src: SRC, tgt: "type[TGT | dict[str, Any]] ", /) -> "TGT | dict[str, Any]":
-#         ...
-
-#     def convert(self, source: SRC, /) -> TGT:
-#         """convert `source` node
-
-#         Args:
-#             source: A bioimageio description node
-
-#         Raises:
-#             ValidationError: conversion failed
-#         """
-#         data = self.convert_as_dict(source)
-#         return assert_all_params_set_explicitly(self.tgt)(**data)
-
-#     def convert_as_dict(self, source: SRC) -> Dict[str, Any]:
-#         return cast(Dict[str, Any], self._convert(source, dict))
-
-
-# A TypeVar bound to a TypedDict seemed like a good way to add converter kwargs:
-# ```
-# class ConverterKwargs(TypedDict):
-#     pass
-# KW = TypeVar("KW", bound=ConverterKwargs, default=ConverterKwargs)
-# ```
-# sadly we cannot use a TypeVar bound to TypedDict and then unpack it in the Converter methods,
-# see https://github.com/python/typing/issues/1399
-# Therefore we use a TypeVarTuple and positional only args instead
-# (We are avoiding ParamSpec for its ambiguity 'args vs kwargs')
-CArgs = TypeVarTuple("CArgs")
-
-
-class Converter(Generic[SRC, TGT, Unpack[CArgs]], ABC):
-    # src: ClassVar[Type[SRC]]
-    # tgt: ClassVar[Type[TGT]]
-    # note: the above is not yet possible, see https://github.com/python/typing/discussions/1424
-    # we therefore use an instance
-    def __init__(self, src: Type[SRC], tgt: Type[TGT], /):
-        super().__init__()
-        self.src: Final[Type[SRC]] = src
-        self.tgt: Final[Type[TGT]] = tgt
-
-    @abstractmethod
-    def _convert(
-        self, src: SRC, tgt: "type[TGT | dict[str, Any]]", /, *args: Unpack[CArgs]
-    ) -> "TGT | dict[str, Any]": ...
-
-    # note: the following is not (yet) allowed, see https://github.com/python/typing/issues/1399
-    #       we therefore use `kwargs` (and not `**kwargs`)
-    # def convert(self, source: SRC, /, **kwargs: Unpack[KW]) -> TGT:
-    def convert(self, source: SRC, /, *args: Unpack[CArgs]) -> TGT:
-        """convert `source` node
-
-        Args:
-            source: A bioimageio description node
-
-        Raises:
-            ValidationError: conversion failed
-        """
-        data = self.convert_as_dict(source, *args)
-        return assert_all_params_set_explicitly(self.tgt)(**data)
-
-    def convert_as_dict(self, source: SRC, /, *args: Unpack[CArgs]) -> Dict[str, Any]:
-        return cast(Dict[str, Any], self._convert(source, dict, *args))
 
 
 class NodeWithExplicitlySetFields(Node):
@@ -350,6 +257,7 @@ class ResourceDescrBase(
 
     @property
     def root(self) -> Union[RootHttpUrl, DirectoryPath, ZipPath]:
+        """The URL/Path prefix to resolve any relative paths with."""
         return self._root
 
     @classmethod
@@ -379,6 +287,7 @@ class ResourceDescrBase(
     def load(
         cls, data: BioimageioYamlContent, context: Optional[ValidationContext] = None
     ) -> Union[Self, InvalidDescr]:
+        """factory method to create a resource description object"""
         context = context or validation_context_var.get()
         assert isinstance(data, dict)
         with context.replace(log_warnings=False):  # don't log warnings to console
